pyproject.toml
[tool.poetry]
name = "gift-card-customizer"
version = "0.1.0"
description = "A web application for creating custom e-gift cards"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" },
    { from = "app", include = "services" }
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
aiofiles = "^23.2.1"
pillow = "^10.2.0"
ollama = "^0.1.6"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
fpdf = "^1.7.2"
qrcode = "^8.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 